Android: split editor and message HTML parsing modes
When we translated HTML to spans in Android, we want the paragraph tags (`<p>`) to translate to a double line break ( `\n\n`) since Android doesn't have support for inter-paragraph spacing and we need to simulate this by using this workaround.

However, this is not something we can support in the actual editor since it would break the index matching, add new points in the text that can be selected, removed, etc. and would make it crash.

So we end up having to differentiate between HTML parsing for the editor and HTML parsing to display in messages to support this change. Also, this means the rendering on the editor and what's rendered later for the sent message might not match 1:1 as until now.